Altered FindItemViewModel to separate to a container view model. Changed that view model classs ot show Selected item on the form. Needs to be able to edit the item after it is selected.
